+++ lib/cisco_node_utils/interface.rb
@@ -96,8 +96,7 @@ module Cisco
     end
 
     def create
-      feature_vlan_set(true) if @name[/vlan/i]
-      feature_vlan_set(true) if @name[/bdi/i]
+      feature_vlan_set(true) if @name[/(vlan|bdi)/i]
       config_set('interface', 'create', name: @name)
     rescue Cisco::CliError
       # Some XR platforms do not support channel-group configuration