+++ CHANGELOG.md
@@ -17,14 +17,13 @@ Changelog
   * bridge_domain_vni (@rkorlepa)
 * Encapsulation Profile
   * vni_encapsulation_profile (@rkorlepa)
-* Interface Bridge Domain
-  * interface bdi (@rkorlepa)
 
 #### NetDev Resources
 *
 
 ### Added
 
+* Added support for bdi interfaces to interface provider.
 * Added a new node util to handle bridge domain range cli for member vni
 * Added Bridge Domain, VNI and encapsulation profile node utils for MT-FULL on Nexus 7k.
 * Added client support for gRPC on IOS XR.